Refactoring of files/code
