+++ nexus-writer/src/nexus_structure/entry/event_data.rs
@@ -128,7 +128,7 @@ impl NexusSchematic for EventData {
 impl NexusMessageHandler<InitialiseNewNexusRun<'_>> for EventData {
     fn handle_message(
         &mut self,
-        InitialiseNewNexusRun(parameters): &InitialiseNewNexusRun<'_>,
+        InitialiseNewNexusRun { parameters }: &InitialiseNewNexusRun<'_>,
     ) -> NexusHDF5Result<()> {
         self.offset = Some(parameters.collect_from);
         self.event_time_zero.add_attribute_to(
